requirements.txt
click==8.1.3
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dateutil==2.8.2
six==1.16.0
